requirements.txt
pyping
spidev
Pillow